--- .storybook/addons.js
@@ -1,2 +1,0 @@
-import '@storybook/addon-knobs/register'
-import '@storybook/addon-storysource/register'
