+++ .storybook/main.js
@@ -1,0 +1,6 @@
+module.exports = {
+  "stories": [
+    "../src/**/*.stories.@(js|jsx|ts|tsx)"
+  ],
+  addons: ['@storybook/addon-storysource', '@storybook/addon-knobs'],
+}
