+++ package.json
@@ -59,9 +59,9 @@
     "@significa/eslint-config": "0.0.8",
     "@significa/prettier-config": "0.0.8",
     "@significa/tsconfig-config": "0.0.8",
-    "@storybook/addon-knobs": "^5.2.8",
-    "@storybook/addon-storysource": "^5.2.8",
-    "@storybook/react": "^5.2.8",
+    "@storybook/addon-knobs": "^6.1.17",
+    "@storybook/addon-storysource": "^6.1.17",
+    "@storybook/react": "^6.1.17",
     "@types/jest": "^24.0.25",
     "@types/react": "^16.9.17",
     "@types/react-dom": "^16.9.4",
@@ -69,6 +69,7 @@
     "@types/react-helmet": "^5.0.14",
     "@types/styled-components": "^4.4.1",
     "babel-loader": "^8.0.5",
+    "babel-preset-react-app": "^10.0.0",
     "cz-conventional-changelog": "^3.0.2",
     "eslint-plugin-jest": "^23.2.0",
     "gatsby-plugin-canonical-urls": "^2.1.18",
